Changed JsonLite version to 1.1.2

+++ jsonlite/src/jsonlite_token_pool.c
@@ -190,7 +190,7 @@ static uint32_t MurmurHash2(const void * key, size_t len)
     
     // Initialize the hash to a 'random' value
     
-    uint32_t h = len;
+    uint32_t h = (uint32_t)len;
     
     // Mix 4 bytes at a time into the hash
